+++ src/ULinkGame.Abstractions/SessionResumeStatus.cs
@@ -1,0 +1,10 @@
+namespace ULinkGame.Abstractions
+{
+    public enum SessionResumeStatus
+    {
+        Resumed,
+        StateRefreshRequired,
+        StateLost,
+        Unauthorized
+    }
+}